support extract synthetics location_failure_conditions
