+++ cmd/get/get_alertsconditions.go
@@ -179,6 +179,17 @@ func GetAllConditionsByAlertPolicyID(id int64) (*newrelic.AlertsConditionList, e
 		}
 	}
 
+	// location_failure_conditions uses different methods for pagination,
+	// if a page requested doesn't exist, it returns the first page instead of empty entires.
+	// use seperate logic to get location_failure_conditions
+	list, resp, err := client.AlertsConditions.List(context.Background(), conditionsOptions, newrelic.ConditionLocation)
+	if err != nil || resp.StatusCode >= 400 {
+		ret := tracker.ToReturnValue(false, tracker.OPERATION_NAME_GET_CONDITIONS_BY_POLICY_ID, fmt.Errorf("%v.Response: %v. Error: %v.", newrelic.ConditionLocation, resp, err), tracker.ERR_REST_CALL, "")
+		return nil, err, ret
+	}
+	allList.AlertsLocationConditionList = &newrelic.AlertsLocationConditionList{}
+	allList.AlertsLocationConditions = append(allList.AlertsLocationConditions, list.AlertsLocationConditions...)
+
 	ret := tracker.ToReturnValue(true, tracker.OPERATION_NAME_GET_CONDITIONS_BY_POLICY_ID, nil, nil, "")
 
 	return allList, err, ret
